Add new test for NoGroupTest
